코드 래빗 수정

+++ src/main/java/com/hrr/backend/domain/challenge/repository/ChallengeStaticsRepository.java
@@ -9,7 +9,7 @@ import com.hrr.backend.domain.challenge.entity.Challenge;
 import com.hrr.backend.domain.challenge.entity.ChallengeStatics;
 import com.hrr.backend.global.common.enums.FavorType;
 
-public interface ChallengeStaticsRepository extends JpaRepository<ChallengeStatics, Integer> {
+public interface ChallengeStaticsRepository extends JpaRepository<ChallengeStatics, Long> {
 
 	// 특정 챌린지에서 특정 선호타입과 값에 해당하는 통계 정보 조회
 	Optional<ChallengeStatics> findByChallengeAndFavorTypeAndFavorValue(Challenge challenge, FavorType type, String value);
